Added the whole Vulnerability Seeker enchantment
